Move commands to folder and plugin.yml and .gitignore edits
