+++ .gitignore
@@ -1,1 +1,1 @@
-toolkitplugin/dependency-reduced-pom.xml
+.vscode/
